Added [ReadOnlyInspector] property
I cleaned some empty lines of code, removed "_isAlreadyAttacked" bool from "NewPlayerData" because the bool "canShoot" already existed and it was still working with it instead of "_isAlreadyAttacked".
And now if you want something to show in the inspector but its not something you want people to be able to change then make it visible in the inspector (public/SerializeField) and add [ReadOnlyInspector] before it.